Osa 4 tehty

+++ osa4/blogilista/controllers/blogs.js
@@ -1,6 +1,6 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const middleware = require('../utils/middleware')
 require('express-async-errors')
 
 blogsRouter.get('/', async (request, response) => {
@@ -8,28 +8,32 @@ blogsRouter.get('/', async (request, response) => {
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
-  const users = await User.find({})
+  const user = request.user
 
   const blog = await new Blog({
     author: body.author,
     title: body.title,
     url: body.url,
     likes: body.likes,
-    user: users[0]._id
+    user: user._id
   })
 
   const savedBlog = await blog.save()
-  users[0].blogs = users[0].blogs.concat(savedBlog._id)
-  await users[0].save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
 
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (!blog || blog.user.toString() === request.user._id.toString()){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+  return response.status(401).json({ error: 'user id does not match blog creator id' })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
